requirement.txt
google-adk
elasticsearch>=8.0.0
python-dotenv>=1.0.0